Split tests into separate files (#66)
